src/requirements.txt
PyQt5>=5.15
rembg>=2.0.50
pillow>=10.0
numpy>=1.24
onnxruntime>=1.16
